alternative_automation/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.4
openpyxl==3.1.2
python-dateutil==2.8.2
workalendar==16.0
numpy==1.26.3
python-dotenv==1.0
httpx==0.25.2